requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.1
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
